Move methods to utils file
